+++ lib/hcloud/entities/prices/floating_ip.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HCloud
+  module Prices
+    class FloatingIP < Entity
+      attribute :price_monthly, :amount
+    end
+  end
+end