+++ lib/hcloud/entities/prices/floating_ips.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module HCloud
+  module Prices
+    class FloatingIPs < Entity
+      attribute :type
+      attribute :prices, :price
+    end
+  end
+end