+++ lib/hcloud/entities/prices/server_backup.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HCloud
+  module Prices
+    class ServerBackup < Entity
+      attribute :percentage
+    end
+  end
+end