+++ lib/hcloud/entities/prices/server_type.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HCloud
+  module Prices
+    class ServerType < Entity
+      attribute :id, :integer
+      attribute :name
+      attribute :prices, :price, array: true, default: -> { [] }
+    end
+  end
+end